Making Preference Screen Fragment and
Implemented the MoviesPreferenceFragment to put ListPreference for the sort by key value selected from an array. The keys and key values shared with the build url and settings preferences MUST be 'translatable="false"' for the strings in strigs.xml

FYI: OnPreferenceChangeListener is used to listen for a change in list preference and also initialize the default value of the key when the app is downloaded. if not ListPreference default value will not change or be initialized