Visual polishing: Loading Indicator
Implemented loading indicator to be visible on pre execute of the asyncTask and hide it after the task is done

FYI: The loading indicator is initially invisible. I also changed to frameLayout so that the loading indicator will appear in the middle. if it was on relative, it will appear at the top left corner of the screen