Broken error fixed
I had to uninstall the app on my device, and install it again, for my database to prepopulate correctly. And then onStartLoading() method of my asyncTask, check if i have data in my databse already to display, if not it will forceload therby making a network call.

Due to i have data pre populated already in database, there will be no forceload