Implemented Movie Reviews
I added a menu item for movie reviews. Passed the movieId as an intent to the review activity when the review menu is clicked from the movie detail activity. I implemented the review recycler view adapter from list item

I will have to implement viewModel to keep data through configuration changes and also liveData to minimize the number of time we query database for data only when data is changed