Implemented the Movie Favourites list page
It was thesame process to make the MainActivity class, just that i used a different Dao query for the FavouriteViewModel. Also i implemented correct onBackPress in movieDetail to keep the state of the recyclerView scroll when we go back to the page. I also implemented a visual polish to know when there is no review of a movie as well as favourite list

Refactored the factory view model class to accomodate favourite view model.